delete button, about page

+++ need/models.py
@@ -4,7 +4,7 @@ from django.utils import timezone
 
 class Need(models.Model):
 	what = models.CharField(max_length=200)
-	amount = models.IntegerField(default=1)
+	amount = models.PositiveSmallIntegerField(default=1)
 	where = models.CharField(max_length=200)
 	#PLACES = (('HWBR', 'HWBR'), ('Emporhalle', 'Emporhalle'), ('Physik', 'Physik'),)
 	#place = models.CharField(max_length=1, choices=PLACES, default='HWBR')
